Move the time functions to a common module
