+++ client/sdl/i_system.h
@@ -56,18 +56,6 @@ void I_Endoom(void);
 // for the zone management.
 void *I_ZoneBase (size_t *size);
 
-
-// returns current time in nanoseconds.
-dtime_t I_GetTime();
-
-dtime_t I_ConvertTimeToMs(dtime_t value);
-dtime_t I_ConvertTimeFromMs(dtime_t value);
-
-// yields to the OS for the specified time (in nanoseconds)
-void I_Sleep(dtime_t);
-// yields to the OS for 1 millisecond
-void I_Yield();
-
 //
 // Called by D_DoomLoop,
 // called before processing each tic in a frame.
@@ -128,9 +116,6 @@ std::string I_ConsoleInput (void);
 
 // Returns true if there will be no application window
 bool I_IsHeadless();
-
-// [RH] Returns millisecond-accurate time
-dtime_t I_MSTime (void);
 
 // [RH] Title string to display at bottom of console during startup
 extern char DoomStartupTitle[256];
