+++ client/sdl/i_video.h
@@ -86,8 +86,6 @@ void I_SetPalette(const argb_t* palette);
 void I_BeginUpdate();
 void I_FinishUpdate();
 
-void I_WaitVBL(int count);
-
 void I_SetWindowCaption(const std::string& caption = "");
 void I_SetWindowIcon();
