+++ client/sdl/i_video_sdl20.cpp
@@ -37,6 +37,7 @@
 
 #include "v_palette.h"
 #include "i_system.h"
+#include "i_time.h"
 #include "i_input.h"
 #include "i_icon.h"
