+++ client/src/st_new.cpp
@@ -41,6 +41,7 @@
 #include "w_wad.h"
 #include "z_zone.h"
 #include "i_system.h"
+#include "i_time.h"
 #include "st_stuff.h"
 #include "hu_drawers.h"
 #include "hu_elements.h"
