+++ server/src/sv_level.cpp
@@ -34,6 +34,7 @@
 #include "gi.h"
 
 #include "i_system.h"
+#include "i_time.h"
 #include "minilzo.h"
 #include "m_random.h"
 #include "p_acs.h"
