+++ server/src/sv_sqpold.cpp
@@ -29,6 +29,7 @@
 #include "d_main.h"
 #include "d_player.h"
 #include "i_system.h"
+#include "i_time.h"
 #include "p_ctf.h"
 #include "g_gametype.h"
